+++ server/src/main/java/com/darly/db/entity/record/RecordMapping.java
@@ -15,14 +15,14 @@ public class RecordMapping {
     private String recordDate;
     private Float recordDistance;
     private Long recordTime;
-    private Float recordPace;
+    private Integer recordPace;
     private Integer recordHeart;
     private Integer recordCalories;
     private String[] coordinateLatitudes;
     private String[] coordinateLongitudes;
 
     @QueryProjection
-    public RecordMapping(Long recordId, String recordTitle, Long recordDate, Float recordDistance, Long recordTime, Float recordPace, Integer recordHeart, Integer recordCalories, String coordinateLatitudes, String coordinateLongitudes) {
+    public RecordMapping(Long recordId, String recordTitle, Long recordDate, Float recordDistance, Long recordTime, Integer recordPace, Integer recordHeart, Integer recordCalories, String coordinateLatitudes, String coordinateLongitudes) {
         this.recordId = recordId;
         this.recordTitle = recordTitle;
         this.recordDate =  new SimpleDateFormat("yyyy/MM/dd").format(new Date(recordDate * 1000));
